# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chess Academy API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterRequest {
  username: string;
  email: string;
  password: string;
  displayName?: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface UserProfile {
  id: string;
  username: string;
  email: string;
  displayName?: string;
  level: string;
  rating: number;
  lessonsCompleted: number;
  openingsStudied: number;
  joinedAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface Opening {
  id: string;
  name: string;
  eco: string;
  description: string;
  moves: string[];
  category: string;
  difficulty: string;
  keyIdeas: string[];
  popularity: number;
}

export interface Exercise {
  id: string;
  type: string;
  question: string;
  options?: string[];
  answer: string;
  explanation: string;
}

export interface Lesson {
  id: string;
  title: string;
  description: string;
  category: string;
  difficulty: string;
  duration: number;
  content: string;
  exercises: Exercise[];
  order: number;
}

export interface UserProgress {
  userId: string;
  completedLessons: string[];
  completedOpenings: string[];
  totalPoints: number;
  streak: number;
  lastActive: string;
}

export interface UpdateProgressRequest {
  lessonId?: string;
  openingId?: string;
  pointsEarned?: number;
}