# lib/api-zod/src/generated/types/exercise.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chess Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface Exercise {
  id: string;
  type: string;
  question: string;
  options?: string[];
  answer: string;
  explanation: string;
}