# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chess Academy API
 * OpenAPI spec version: 0.1.0
 */
import type { Exercise } from "./exercise";

export interface Lesson {
  id: string;
  title: string;
  description: string;
  category: string;
  difficulty: string;
  duration: number;
  content: string;
  exercises: Exercise[];
  order: number;
}