# lib/api-zod/src/generated/types/opening.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chess Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface Opening {
  id: string;
  name: string;
  eco: string;
  description: string;
  moves: string[];
  category: string;
  difficulty: string;
  keyIdeas: string[];
  popularity: number;
}