# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chess Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterRequest {
  username: string;
  email: string;
  password: string;
  displayName?: string;
}