# lib/api-zod/src/generated/types/updateProgressRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chess Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProgressRequest {
  lessonId?: string;
  openingId?: string;
  pointsEarned?: number;
}