# lib/api-zod/src/generated/types/userProgress.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Chess Academy API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProgress {
  userId: string;
  completedLessons: string[];
  completedOpenings: string[];
  totalPoints: number;
  streak: number;
  lastActive: string;
}